rearrangement of pages
